Moved base controllers into the controllers folder and namespace
Instead of extending:
Twist\Core\Classes\BaseController
Now you just need to extend:
Base

The same applies for BaseUser and BaseAJAX Controllers
