+++ src/twist/Core/Controllers/Base.controller.php
@@ -21,7 +21,7 @@
 	 *
 	 */
 
-	namespace Twist\Core\Classes;
+	namespace Twist\Core\Controllers;
 	use Twist\Core\Models\Route\Meta;
 	use Twist\Core\Classes\Error;
 
@@ -29,7 +29,7 @@
 	 * Base Controller should be used as an extension to every controller class that is made with the exception of those controllers that use BaseControllerAJAX and BaseControllerUser.
 	 * @package Twist\Core\Classes
 	 */
-	class BaseController{
+	class Base{
 
 		protected $arrMessages = array();
 		protected $arrAliasURIs = array();
